API Documentation

Functions

Base.size — Method

size(A::MultiAddInflation)

Return the dimension of the additive inflation of A.

RecipesBase.apply_recipe — Method

    heatmap(M::HermiteMap; start::Int64=1, color, degree)

Plot recipe for an ExpandedFunction. We can either plot the number of occurences of each variable (columns) in each map component (rows) if degree = false (default behavior), or the maximum multi-index of the features identified for each variable (columns) in each map component (rows) if degree = true.

RecipesBase.apply_recipe — Method

    heatmap(M::HermiteMap; start::Int64=1, color, degree)

Plot recipe for an HermiteMap. We can either plot the number of occurences of each variable (columns) in each map component (rows) if degree = false (default behavior), or the maximum multi-index of the features identified for each variable (columns) in each map component (rows) if degree = true.

TransportBasedInference.DPhyPolyHermite — Method

TransportBasedInference.Diff — Method

Diff(f,k)

By recurrence, create the k-th derivative of the scalar function f, based on ForwardDiff.

TransportBasedInference.Diff — Method

Diff(f)

Create the derivative of the scalar function f, based on ForwardDiff.

TransportBasedInference.allequal — Method

allequal(x, r)

A function to verify that all the entries of the vector x are equal to r

TransportBasedInference.bisection — Method

Implement the classical Bisection algorithm on the segment [xm, xp] of the real function f.

TransportBasedInference.clenshaw_curtis — Method

clenshaw_curtis(N)

Compute the nodes x and weights w for integrating a continuous functions from [-1,1] using the Clenshaw-Curtis integration rule with order N.

TransportBasedInference.csne — Method

Solve the corrected semi-normal equations R'Rx=A'b.

x, r = csne(R, A, b) solves the least-squares problem

minimize ||r||_2, where r := b - A*x

using the corrected semi-normal equation approach described by Bjork (1987). Assumes that R is upper triangular.

TransportBasedInference.degree — Method

    degree(P)

Return the degree of the polynomial P

TransportBasedInference.elementproductmatmul! — Method

elementproductmatmul!(d, A, B, c)

Compute in-place the product (A ∘ B)*c, where ∘ denotes the element-wise product of two matrices.

TransportBasedInference.exactn — Method

    exactn(N)

A function to create a 1D sample with exactly mean 0 and covariance 1 (The samples are no longer i.i.d but this can be usueful when the initialization of the problem is challenging.)

TransportBasedInference.fact2 — Method

fact2(n)

Compute the double factorial of n, i.e. the product of the odd (if n is odd) or even (if n is even) numbers up to n.

TransportBasedInference.grad_x_logeval — Method

Compute g′(x)/g(x) i.e d/dx log(g(x))

TransportBasedInference.gradlog_pdf — Method

gradlog_pdf(x)

Compute the gradient of the log of the density of the univariate standard Gaussian distribution (zero mean and unitary standard deviation).

TransportBasedInference.greedyfit — Method

greedyfit(m::Int64, Nx::Int64, X, Xvalid, maxterms::Int64; withconstant, withqr, maxpatience, verbose, hessprecond) -> Tuple{HermiteMapComponent,Tuple{Array{Float64,1},Array{Float64,1}}}

An adaptive routine to estimate a sparse approximation of an HermiteMapComponent based on the pair of ensemble matrices X (training set) and Xvalid (validation set).

TransportBasedInference.greedyfit — Method

greedyfit(m::Int64, Nx::Int64, X, maxterms::Int64; withconstant, withqr, maxpatience, verbose, hessprecond) -> Tuple{HermiteMapComponent,Array{Float64,1}}

An adaptive routine to estimate a sparse approximation of an HermiteMapComponent based on the ensemble matrix X.

TransportBasedInference.greedyfit — Method

greedyfit(Nx::Int64, poff::Int64, pdiag::Int64, X::AbstractArray{Float64,2}, Xvalid::AbstractArray{Float64,2}, maxfamiliesoff::Int64, λ::Float64, δ::Float64, γ::Float64; maxpatience, verbose) -> Tuple{SparseRadialMapComponent,Tuple{Array{Float64,1},Array{Float64,1}}}

An adaptive routine to estimate a sparse approximation of an SparseRadialMapComponent based on the pair of ensemble matrices X (training set) and Xvalid (validation set).

TransportBasedInference.greedyfit — Method

greedyfit(Nx::Int64, poff::Int64, pdiag::Int64, X::AbstractArray{Float64,2}, maxfamiliesoff::Int64, λ::Float64, δ::Float64, γ::Float64; maxpatience, verbose) -> Tuple{SparseRadialMapComponent,Array{Float64,1}}

An adaptive routine to estimate a sparse approximation of an SparseRadialMapComponent based on the ensemble matrix X.

TransportBasedInference.hesslog_pdf — Method

hesslog_pdf(x)

Compute the hessian of the log of the density of the univariate standard Gaussian distribution (zero mean and unitary standard deviation).

TransportBasedInference.log_pdf — Method

log_pdf(x)

Compute the log of the density of the univariate standard Gaussian distribution (zero mean and unitary standard deviation).

TransportBasedInference.log_pdf_banana — Method

log_pdf_banana(X; μ = 0.0, σ = 2.0, bananicity = 0.2)

Compute the logarithm of the porbability density function (pdf) of the Banana distribution at X. See documentation of sample_banana for details on the Banana distribution.

TransportBasedInference.lorenz63! — Method

lorenz63!(du,u,p,t)

Compute in-place the right-hand-side of the Lorenz-63 system for a state u at time t, and store it in du. p is vector of user-defined parameters.

TransportBasedInference.lorenz96! — Method

lorenz96!(du,u,p,t)

Compute in-place the right-hand-side of the Lorenz-96 system for a state u at time t, and store it in du. p is vector of user-defined parameters.

TransportBasedInference.mean_hist — Method

mean_hist(hist)

Stack together the mean of the different ensembles.

TransportBasedInference.meas — Method

meas(X, Ny, Nx)

A function to extract the first Ny lines of the ensemble matrix X, typically storing the observations.

TransportBasedInference.member — Method

member(X, idx)

Extract the idx ensemble member of the ensemble matrix X

TransportBasedInference.observe — Method

observe(h::Function, X, t::Float64, Ny::Int64, Nx::Int64; P)

Evaluate the function h for the different state vectors of the X at time t, and store the results in the first Ny columns of X. X is an ensemble matrix that contains the observation vectors in the first Ny lines, and the state vectors in the lines Ny+1 to Ny+Nx. The code can run in serial or with multithreading.

TransportBasedInference.observe — Method

observe(F::StateSpace, x::Array{Float64,1}, t::Float64) -> Any

Apply the observation operator of the StateSpace F to the ensemble matrix X at time t.

TransportBasedInference.qraddrow — Method

Add a row and update a Q-less QR factorization.

qraddrow!(R, a) returns the triangular part of a QR factorization of [A; a], where A = QR for some Q. The argument 'a' should be a row vector.

TransportBasedInference.qrdelcol — Method

Delete the k-th column and update a Q-less QR factorization.

R = qrdelcol(R,k) deletes the k-th column of the upper-triangular R and restores it to upper-triangular form. On input, R is an n x n upper-triangular matrix. On output, R is an n-1 x n-1 upper triangle.

18 Jun 2007: First version of QRdelcol.m.
             Michael Friedlander (mpf@cs.ubc.ca) and
             Michael Saunders (saunders@stanford.edu)
             To allow for R being sparse,
             we eliminate the k-th row of the usual
             Hessenberg matrix rather than its subdiagonals,
             as in Reid's Bartel-Golub LU update and also
             the Forrest-Tomlin update.
             (But Matlab will still be pretty inefficient.)
18 Jun 2007: R is now the exact size on entry and exit.
30 Dec 2015: Translate to Julia.

TransportBasedInference.rdnortho — Method

rdnortho(N::Int64) -> Array{Float64,2}

A routine to generate mean-preserving random rotations.

References:

Nerger, L., Janjić, T., Schröter, J., & Hiller, W. (2012). A unification of ensemble square root Kalman filters. Monthly Weather Review, 140(7), 2335-2345. Tödter, J., & Ahrens, B. (2015). A second-order exact ensemble square root filter for nonlinear data assimilation. Monthly Weather Review, 143(4), 1347-1367.

TransportBasedInference.rmse — Method

rmse(X)

Compute the root-mean square error of the ensemble matrix X

TransportBasedInference.sample_banana — Method

sample_banana(N; μ = 0.0, σ = 2.0, bananicity = 0.2)

Generate N samples [x₁; x₂] of the Banana distribution defined as: x₁ ∼ N(μ, σ²), ϵ ∼ N(0, 1), x₂ ∼ bananicity × (x₁ - σ²) + ϵ.

TransportBasedInference.seqassim — Method

seqassim(F::StateSpace, data::SyntheticData, J::Int64, ϵx::InflationType, algo::SeqFilter, X, Ny, Nx, t0::Float64, Loc::Localization)

Generic API for localized sequential data assimilation for any sequential filter of parent type SeqFilter with localization scheme Loc.

TransportBasedInference.seqassim — Method

seqassim(F::StateSpace, data::SyntheticData, J::Int64, ϵx::InflationType, algo::SeqFilter, X, Ny, Nx, t0::Float64)

Generic API for sequential data assimilation for any sequential filter of parent type SeqFilter.

TransportBasedInference.spread — Method

spread(X)

Compute the spread of the ensemble matrix X

TransportBasedInference.state — Method

state(X, Ny, Nx)

A function to extract the lines Ny+1 to Ny+Nx of the ensemble matrix X, typically storing the state.

TransportBasedInference.vander! — Method

vander!(dV, B::CstLinPhyHermite, maxi::Int64, k::Int64, x)

Compute the Vandermonde matrix for the vector x

TransportBasedInference.vander! — Method

vander!(dV, B::CstLinProHermite, maxi::Int64, k::Int64, x)

Compute the Vandermonde matrix for the vector x

TransportBasedInference.vander! — Method

vander!(dV, B::CstPhyHermite, maxi::Int64, k::Int64, x)

Compute the Vandermonde matrix for the vector x

TransportBasedInference.viewmeas — Method

viewmeas(X, Ny, Nx)

Create a view of the first Ny lines of the ensemble matrix X, typically storing the observations.

TransportBasedInference.viewstate — Method

viewstate(X, Ny, Nx)

Create a view of the lines Ny+1 to Ny+Nx of the ensemble matrix X, typically storing the state.

Types

TransportBasedInference.AdditiveInflation — Method

    (A::AdditiveInflation)(X, start::Int64, final::Int64)

Apply the additive inflation A to the lines start to final of an ensemble matrix X, i.e. xⁱ -> xⁱ + ϵⁱ with ϵⁱ ∼ A.α.

TransportBasedInference.AdditiveInflation — Method

    (A::AdditiveInflation)(X)

Apply the additive inflation A to an ensemble matrix X, i.e. xⁱ -> xⁱ + ϵⁱ with ϵⁱ ∼ A.α.

TransportBasedInference.AdditiveInflation — Method

    (A::AdditiveInflation)(x::Array{Float64,1})

Apply the additive inflation A to the vector x, i.e. x -> x + ϵ with ϵ ∼ A.α.

TransportBasedInference.Basis — Type

abstract type Basis

A structure to hold a basis of functions

Fields

Constructors

Basis(m)

Construct new basis of features

A specific feature basis MyBasis must be a subtype of Basis, and the following routines must be implemented:

Base.show(io::IO, B::MyBasis)` (optional, but desired)

@propagate_inbounds Base.getindex(B::MyBasis, i::Int64)

vander!(dV, B::MyBasis, maxi::Int64, k::Int64, x)

TransportBasedInference.CstLinPhyHermite — Type

struct CstLinPhyHermite <: Basis

The basis composed of the constant function, the identity, and physicist Hermite functions

Fields

m::Int64

TransportBasedInference.CstLinProHermite — Type

struct CstLinProHermite <: Basis

The basis composed of the constant function, the identity, and probabilistic Hermite functions

Fields

m::Int64

TransportBasedInference.CstPhyHermite — Type

struct CstPhyHermite <: Basis

The basis composed of the constant function and physicist Hermite functions

Fields

m::Int64

TransportBasedInference.CstProHermite — Type

struct CstProHermite <: Basis

The basis composed of the constant function and probabilistic Hermite functions

Fields

m::Int64

TransportBasedInference.ExpandedFunction — Type

struct ExpandedFunction

ExpandedFunction decomposes a multi-dimensional function f:R^{Nx} → R onto a basis of MultiFunction ψα where cα are scalar coefficients for each MultiFunction: f(x1, x2, ..., xNx) = ∑α cα ψ_α(x1, x2, ..., xNx), where Nψ is the number of MultiFunctions used, and Nx is the dimension of the input vector x.

Fields

m::Int64

Nψ::Int64

Nx::Int64

B::MultiBasis

idx::Array{Int64,2}

dim::Array{Int64,1}

coeff::Array{Float64,1}

TransportBasedInference.Hermite — Type

Hermite

An abstract type for 1D Hermite polynomials and functions

TransportBasedInference.HermiteMap — Type

struct HermiteMap

Fields

m::Int64

Nx::Int64

L::LinearTransform

C::Array{HermiteMapComponent,1}

TransportBasedInference.HermiteMapComponent — Type

struct HermiteMapComponent

Fields

m::Int64

Nψ::Int64

Nx::Int64

I::IntegratedFunction

α::Float64

TransportBasedInference.IdFilter — Type

IdFilter <: SeqFilter

An immutable structure for the identity filter.

Fields

Δtdyn::Float64: time-step of the dynamical model

Δtobs::Float64: time-step between two observations

Constructors

IdFilter(Δtdyn, Δtobs): set up an identity filter

TransportBasedInference.IdFilter — Method

(filter::IdFilter)(X, ystar, t)

Applies the identity transformation to the ensemble `X`.

TransportBasedInference.IdentityInflation — Type

IdentityInflation <: InflationType

An type to store identity inflation :

Define additive inflation: x <- x

TransportBasedInference.IdentityInflation — Method

    (A::IdentityInflation)(X)

Apply an IdentityInflation A on an ensemble matrix X, i.e. xⁱ -> xⁱ

TransportBasedInference.InflationType — Type

InflationType

An abstract type for Inflation.

TransportBasedInference.IntegratedFunction — Type

struct IntegratedFunction

An immutable structure to hold the integrated function with g a Rectifier and f an ExpandedFunction.

Fields

m::Int64

Nψ::Int64

Nx::Int64

g::Rectifier

f::ExpandedFunction

TransportBasedInference.InvPreconditioner — Type

InvPreconditioner

An immutable structure to hold the inverse of the preconditioner. For instance, this structure can be used to hold the estimate of the inverse of the Hessian in the BFGS algorithm.

TransportBasedInference.LinHermiteMapComponent — Type

struct LinHermiteMapComponent

LinHermiteMapComponent is a type to hold the composition of a LinearTransform: a linear transformation that rescales the samples to get zero mean and unitary standard deviation component-wise and an HermiteMapComponent element.

Fields

L::LinearTransform

C::HermiteMapComponent

TransportBasedInference.LinearTransform — Type

LinearTransform

TransportBasedInference.Model — Type

    Model

A structure to perform the twin-experiment

Fields

Nx : Dimension of the state variable

Ny : Dimension of the observation variable

Δtdyn : Time-step for the dynamical model

Δtobs : Time step between two observations of the state

ϵx : Process noise

ϵy : Observation noise

π0 : Multivariate distribution for the initial condition

Tburn : Number of steps to burn from the end of the spin up to compute the metrics

Tstep : Number of steps for which the filter to be tested is applied

Tspinup : Number of steps of spin-up phase, i.e. number of steps to generate the initial ensemble for the filtering algorithms

F : State-Space Model

TransportBasedInference.MultiAddInflation — Type

MultiAddInflation <: InflationType

An type to store multiplico-additive inflation :

Define multiplico-additive inflation: xⁱ -> x̄ + β*(xⁱ - x̄) + ϵⁱ with ϵⁱ ∼ α and β a scalar

Fields:

Nx : dimension of the vector

'β' : Multiplicative inflation factor

'α' : Distribution of the additive inflation

Constructors:

MultiAddInflation(Nx::Int64, β::Real, α::ContinuousMultivariateDistribution)

MultiAddInflation(Nx::Int64, β::Float64, m::Array{Float64,1}, Σ)

MultiAddInflation(Nx::Int64, β::Float64, m::Array{Float64,1}, σ::Array{Float64,1})

MultiAddInflation(Nx::Int64, β::Float64, m::Array{Float64,1}, σ::Float64)

TransportBasedInference.MultiAddInflation — Method

    (A::MultiAddInflation)(X, start::Int64, final::Int64)

Apply the multiplicat inflation A to the lines start to final of an ensemble matrix X, i.e. xⁱ -> x̄ + β*(xⁱ - x̄) + ϵⁱ with ϵⁱ ∼ A.α and β a scalar, usually ∼ 1.0.

TransportBasedInference.MultiAddInflation — Method

    (A::MultiAddInflation)(X, start::Int64, final::Int64)

Apply the multiplico-additive inflation A to the ensemble matrix X, i.e. xⁱ -> x̄ + β*(xⁱ - x̄) + ϵⁱ with ϵⁱ ∼ A.α and β a scalar, usually ∼ 1.0.

TransportBasedInference.MultiBasis — Type

struct MultiBasis

An immutable structure to hold the Nx-cartesian product of the 1D basis of functions B MultiB = B × B× ... × B (Nx elements). The basis B contains m elements.

Fields

B::Basis

Nx::Int64

TransportBasedInference.MultiFunction — Type

struct MultiFunction

An immutable structure to hold an elementary function F: Rᵏ → R that can be decomposed as the product of univariate basis, where each basis is finite and contains constant and/or linear and Hermite functions F(x1, x2, ..., xk) = f1(x1) × f2(x2) × ... × fM(x_k)

Fields

m::Int64

Nx::Int64

B::MultiBasis

α::Array{Int64,1}

TransportBasedInference.MultiplicativeInflation — Type

MultiplicativeInflation <: InflationType

An type to store multiplicative inflation :

xⁱ -> xⁱ + β*(xⁱ - x̄) with β a scalar

Fields:

'β' : multiplicative inflation factor

TransportBasedInference.MultiplicativeInflation — Method

    (A::MultiplicativeInflation)(X, start::Int64, final::Int64)

Apply the multiplicative inflation A to the lines start to final of an ensemble matrix X, i.e. xⁱ -> x̄ + β*(xⁱ - x̄) with β scalar, usually ∼ 1.0.

TransportBasedInference.MultiplicativeInflation — Method

    (A::MultiplicativeInflation)(X)

Apply the multiplicative inflation A to an ensemble matrix X, i.e. xⁱ -> x̄ + β*(xⁱ - x̄) with β scalar, usually ∼ 1.0.

TransportBasedInference.Parallel — Type

Parallel

An abstract type for the different kinds of parallel programming supported.

TransportBasedInference.PhyHermite — Type

PhyHermite <: Hermite

An immutable structure for physicist Hermite functions defined as ψm(x) = Hm(x)*exp(-x^2/2).

Fields

m : order of the function

Poly : physicist Hermite polynomial of order m

scaled : with rescaling to have unitary norm

Constructors

PhyHermite(m, Poly, scaled)

PhyHermite(m; scaled = false)`

TransportBasedInference.PhyPolyHermite — Type

PhyPolyHermite <: Hermite

An immutable structure for physicist Hermite polynomials

Fields

m : order of the polynomial

P : physicist Hermite polynomial of order m

scaled : with rescaling to have unitary norm

TransportBasedInference.ProHermite — Type

ProHermite <: Hermite

An immutable structure for probabilistic Hermite functions defined as ψem(x) = Hem(x)*exp(-x^2/4).

Fields

m : order of the function

Poly : probabilistic Hermite polynomial of order m

scaled : with rescaling to have unitary norm

Constructors

ProHermite(m, Poly, scaled)

ProHermite(m; scaled = false)

TransportBasedInference.ProPolyHermite — Type

ProPolyHermite <: Hermite

An immutable structure for probabilistic Hermite polynomials

Fields

m : order of the polynomial

P : probabilistic Hermite polynomial of order m

scaled : with rescaling to have unitary norm

TransportBasedInference.Rectifier — Type

struct Rectifier

This structure defines a continuous rectifier g, i.e. a positive and monotonically increasing function (e.g. square function, exponential , softplus, explinearunit).

Fields

T::String

TransportBasedInference.SeqFilter — Type

SeqFilter

An abstract type for the ensemble filtering algorithms.

TransportBasedInference.Serial — Type

Serial <: Parallel

A type for serial computing

TransportBasedInference.StateSpace — Type

struct StateSpace

An immutable structure representing the dynamical f and observation h operators. The dynamical model is provided by the right hand side of the ODE to solve. For a system of ODEs, we will prefer an in-place syntax f(du, u, p, t), where p are parameters of the model. We rely on OrdinaryDiffEq to integrate the dynamical system with the Tsitouras 5/4 Runge-Kutta method adaptive time marching. h must be a function of the form h(u, t), where u is the state vector and t is the time.

Fields

f::Function

Propagatation f

h::Function

Observation h

TransportBasedInference.Storage — Type

struct Storage

An immutable structure to hold the evaluation of basis functions

Fields

m::Int64

Nψ::Int64

Nx::Int64

f::ExpandedFunction

ψoff::Array{Float64,2}

ψoffψd::Array{Float64,2}

ψoffψd0::Array{Float64,2}

ψoffdψxd::Array{Float64,2}

ψnorm::Array{Float64,1}

cache_dcψxdt::Array{Float64,2}

cache_gradxd::Array{Float64,2}

cache_dψxd::Array{Float64,1}

cache_integral::Array{Float64,1}

cache_g::Array{Float64,1}

TransportBasedInference.SyntheticData — Type

SyntheticData

A structure to store the synthetic data in a twin-experiment

Fields

tt : time history

x0 : the initial condition

xt : history of the state

yt : history of the observations

TransportBasedInference.Thread — Type

Thread <: Parallel

A type for multi-threaded computing

Index

TransportBasedInference.AdditiveInflation

TransportBasedInference.AdditiveInflation

TransportBasedInference.AdditiveInflation

TransportBasedInference.Basis

TransportBasedInference.CstLinPhyHermite

TransportBasedInference.CstLinProHermite

TransportBasedInference.CstPhyHermite

TransportBasedInference.CstProHermite

TransportBasedInference.ExpandedFunction

TransportBasedInference.Hermite

TransportBasedInference.HermiteMap

TransportBasedInference.HermiteMapComponent

TransportBasedInference.IdFilter

TransportBasedInference.IdFilter

TransportBasedInference.IdentityInflation

TransportBasedInference.IdentityInflation

TransportBasedInference.InflationType

TransportBasedInference.IntegratedFunction

TransportBasedInference.InvPreconditioner

TransportBasedInference.LinHermiteMapComponent

TransportBasedInference.LinearTransform

TransportBasedInference.Model

TransportBasedInference.MultiAddInflation

TransportBasedInference.MultiAddInflation

TransportBasedInference.MultiAddInflation

TransportBasedInference.MultiBasis

TransportBasedInference.MultiFunction

TransportBasedInference.MultiplicativeInflation

TransportBasedInference.MultiplicativeInflation

TransportBasedInference.MultiplicativeInflation

TransportBasedInference.Parallel

TransportBasedInference.PhyHermite

TransportBasedInference.PhyPolyHermite

TransportBasedInference.ProHermite

TransportBasedInference.ProPolyHermite

TransportBasedInference.Rectifier

TransportBasedInference.SeqFilter

TransportBasedInference.Serial

TransportBasedInference.StateSpace

TransportBasedInference.Storage

TransportBasedInference.SyntheticData

TransportBasedInference.Thread

Base.size

RecipesBase.apply_recipe

RecipesBase.apply_recipe

TransportBasedInference.DPhyPolyHermite

TransportBasedInference.Diff

TransportBasedInference.Diff

TransportBasedInference.allequal

TransportBasedInference.bisection

TransportBasedInference.clenshaw_curtis

TransportBasedInference.csne

TransportBasedInference.degree

TransportBasedInference.elementproductmatmul!

TransportBasedInference.exactn

TransportBasedInference.fact2

TransportBasedInference.grad_x_logeval

TransportBasedInference.gradlog_pdf

TransportBasedInference.greedyfit

TransportBasedInference.greedyfit

TransportBasedInference.greedyfit

TransportBasedInference.greedyfit

TransportBasedInference.hesslog_pdf

TransportBasedInference.log_pdf

TransportBasedInference.log_pdf_banana

TransportBasedInference.lorenz63!

TransportBasedInference.lorenz96!

TransportBasedInference.mean_hist

TransportBasedInference.meas

TransportBasedInference.member

TransportBasedInference.observe

TransportBasedInference.observe

TransportBasedInference.qraddrow

TransportBasedInference.qrdelcol

TransportBasedInference.rdnortho

TransportBasedInference.rmse

TransportBasedInference.sample_banana

TransportBasedInference.seqassim

TransportBasedInference.seqassim

TransportBasedInference.spread

TransportBasedInference.state

TransportBasedInference.vander!

TransportBasedInference.vander!

TransportBasedInference.vander!

TransportBasedInference.viewmeas

TransportBasedInference.viewstate